Introduce encrypted_data, move segments, fixing header function names reflect the builder pattern instead of hardcoding header_* and the like
